LPS-55608 Move
